+++ Changelog.md
@@ -3,3 +3,4 @@
 ### v0.1.3
 
 - Checkbox props onChange new second parameter `event: React.ChangeEvent<HTMLInputElement>`
+- new Component `Toast`